simplify flow of ExecuteCommand.run by inlining some old indirections
