+++ Sources/md-babel/Execute/ExecuteCommand.swift
@@ -85,7 +85,14 @@ struct ExecuteCommand: AsyncParsableCommand {
 			FileHandle.standardOutput.write(try JSON.object([:]).data())
 			return
 		}
-		let execute = try Execute(executableContext: context, registry: evaluatorRegistry())
+		let evaluator = Evaluator(
+			configuration: try evaluatorRegistry().configuration(forCodeBlock: context.codeBlock),
+			generateImageFileURL: GenerateImageFileURL(
+				outputDirectory: try outputDirectory(),
+				fileExtension: "svg"  // TODO: Make file extension configurable in converter https://github.com/md-babel/swift-markdown-babel/issues/20
+			)
+		)
+		let execute = Execute(executableContext: context, evaluator: evaluator)
 		let response = await execute()
 
 		try perform(sideEffect: response.executionResult.output?.sideEffect)
@@ -94,6 +101,40 @@ struct ExecuteCommand: AsyncParsableCommand {
 			try json(response, originalLocation: location)
 			.data(formatting: [.prettyPrinted, .sortedKeys, .withoutEscapingSlashes])
 		FileHandle.standardOutput.write(data)
+	}
+}
+
+extension ExecuteCommand {
+	func outputDirectory(
+		outputPath: String? = nil
+	) throws(EvaluatorRegistryFailure) -> URL {
+		guard let outputPath else {
+			return try makeTemporaryOutputDirectory()
+		}
+		return try directory(forPath: outputPath)
+	}
+
+	private func makeTemporaryOutputDirectory() throws(EvaluatorRegistryFailure) -> URL {
+		let tmpDir = FileManager.default.temporaryDirectory
+
+		do {
+			try FileManager.default.createDirectory(at: tmpDir, withIntermediateDirectories: true)
+			return tmpDir
+		} catch let error {
+			throw .directoryLookupFailed("Could not create temporary directory at “\(tmpDir)”: \(error)")
+		}
+	}
+
+	private func directory(forPath path: String) throws(EvaluatorRegistryFailure) -> URL {
+		var isDirectory: ObjCBool = false
+		let pathExists = FileManager.default.fileExists(atPath: path, isDirectory: &isDirectory)
+		guard pathExists else {
+			throw .directoryLookupFailed("Output directory at “\(path)” does not exist")
+		}
+		guard isDirectory.boolValue == true else {
+			throw .directoryLookupFailed("Output path at “\(path)” is not a directory")
+		}
+		return URL(fileURLWithPath: path, isDirectory: true)
 	}
 }
